pkg/preflight/checkers: use Checker interface from preflight package + alloc properly the checkers array
The package was not using the interface from the preflight package. We use it now and we deleted the interface in the checkers package.
Multiple re-allocation was occuring when appending checkers to the array. Now, allocation is done at the beginning by calculating the total number of checkers we will store in the array.