Implement bidirectional cp using sftp
Implement bidirectional copy from host to VM and VM to host using sftp.
The copy command syntax is similar to `docker cp` with source and
destination that can have VM reference name or ID separated by a
filepath using ":". VM reference in copy source means to copy from VM to
host and VM reference in copy destination means to copy from host to VM.

File permissions and owners are also applied to the copied files.
Symlinks are followed and the destination files are copied.