Fix copy when destination is an existing dir
When source is a file and destination is a directory, append the base
file name of source to the destination path, moving the source into
destination directory.
When source is a directory and destination is also a directory with a
different name, copy the source to destination as a subdirectory. If the
destination directory has a file with the same name as the source, fail
to copy. Directory should not overwrite an existing file.