Tiny image import fix
When an image is very small, filesystem creation using mkfs fails due to
not enough space. This adds a minimum base image size that's used when
the image is too small.

Also, fixes the /etc/resolv.conf symlink error when there's no /etc
directory in the image filesystem. /etc dir is created as part of the
import step if it doesn't exists.

Adds an e2e test for this fix using hello-world docker container image
which is tiny and lacks /etc dir.