Add ignite configuration file support
Defines APIs for ignite configurations with `Configuration` object.

Adds a new persistent flag at the root command, `ignite-config`.
This flag can be used to set the file path of an ignite config file.

An ignite config file contains the ignite Configuration object. It can
be used to set the default configurations of various ignite components.
By default, if a config file is found in /etc/ignite/config.yaml, it is
used to configure ignite. `--ignite-config` flag can be used to override
any default global config.