Ignore docker source cleanup err when resource not found
Fixes the snapshot not found error seen in CI when importing VM and
kernel images and running a VM for the first time. The issue was due to
a missing containerd snapshot key which was attempted to be cleaned up.
Since the snapshot key didn't exist, the cleanup returned `snapshot <id>
does not exist: not found` error.
The fix checks if the error is a containerd "not found" error and
ignores it. Any other error during the cleanup is returned as before.