Add support for ignite configuration file
This adds a new API type `Configuration` which is used to define the
default configurations of ignite. A Configuration object can be defined
in a config file as:

```yaml
apiVersion: ignite.weave.works/v1alpha3
kind: Configuration
metadata:
  name: test-config
spec:
  runtime: containerd
  networkPlugin: cni
  vmDefaults:
    memory: "2GB"
    diskSize: "3GB"
    cpus: 2
    sandbox:
      oci: weaveworks/ignite:dev
    kernel:
      oci: weaveworks/ignite-kernel:4.19.47
    ssh: true
```

By default, ignite looks for the config file at /etc/ignite/config.yaml.
If this file is found, ignite uses it to configure the ignite defaults.
`--ignite-config` flag can be used to explicitly pass a configuration
file.