Add ignite e2e test helper package
The helper package provides a type Command with attributes about ignite
execution and methods to execute ignite commands with error checks.
This simplifies a lot of the duplicate code in the e2e tests and
provides reusable code to make e2e tests less verbose.
This Command can be used to run commands for ignite and ignited.