Handle VM start after stop.
At VM stop, the runtime and network info are removed from the VM object.
Attempting to start a VM fails due to unknown runtime and network.
This adds runtime and network-plugin flag to VM start command to
explicitly set the providers if needed. This also causes the VM to use
the default providers when no providers are defined in the VM object.

Update VM start docs.