Add ignited gitops mode e2e test
This tests the basic ignited gitops setup. It creates a bare git repo,
clones it, write a VM config to it, commits and pushes to the bare git
repo. It then runs ignited in gitops mode targeting the bare git repo.
ignited detects the VM config and provisions a VM.