Add github action to update go dependencies
Run `make go-autogen` and create a PR for any updated files.
Schedule to run daily and also enable an option to trigger it manually
on demand for creating PR that updates go dependencies.
This uses branch `go-deps-patch` to create the PR and automatically
deletes it once the PR is closed or merged.