Add default providers at VM start
When a running VM is stopped, the runtime and network info are removed
from the VM object status. Starting the VM without any explicit provider
flags fails due to lack of provider info.
This change adds the default providers in the VM status if the runtime
and network are empty when attempting to start the VM. Any explicitly
passed provider flags continue to override the VM providers.