Assign a random MAC address to the bridge
If we don't do this it will adopt the lowest address of an attached
device, hence change over time.

Remove the previous work-around of getting and setting the random MAC
that Linux assigned, also getting and setting the TUN address since
that doesn't change if you set the bridge address.