ps: Avoid error when the storage is uninitialized
Running ignite ps for the first time on a new machine fails with:
```
FATA[0000] open /var/lib/firecracker/vm: no such file or directory
```

This change handles the error returned when the backend storage is
uninitialized. NewPsOptions() returns an empty list of VMs and allows
the ps execution to continue without any error.