rm: Handle error when container not found
When removing a VM with force, if the container runtime returns error
when it can't find the associated VM container, the force remove fails.
With this change, ignite will checks if the container runtime error is
about container not found and return without any error, no-op, letting
the force remove succeed.
The behavior remains the same for non-forced remove.