Change the log level in RemoveContainer to debug
RemoveContainer() is called before a VM is created and after a VM is
deleted. This results in warning logs with container or task not found
for the basic operations even when nothing wrong happened. The warning
could be misleading. Move the `not found` logs in RemoveContainer() to
debug level to avoid such confusion.