Add loading credentials from docker cli config
This adds a package pkg/runtime/auth with helpers to read the docker cli
config and load the credentials for the host name of a given image. This
is based on nerdctl's dockerconfigresolver.

When using containerd as the runtime, the credentials from docker cli
config is loaded into a containerd remote resolver and passed to the
containerd remote option used for pulling.

When using docker as the runtime, the credentials from docker cli config
is loaded into the docker image pull options in the required format.